Place selected and checklists in NavContext
