+++ client/src/Contexts.ts
@@ -1,0 +1,9 @@
+import { createContext } from "react";
+import { Checklist } from "@/types/Checklist";
+
+export const NavContext = createContext({
+  selected: "",
+  setSelected: (prev: string) => {},
+  checklists: Array<Checklist>(),
+  setChecklists: (prev: any) => {},
+});
