+++ client/src/components/layout/TripLayout.tsx
@@ -1,23 +1,37 @@
 import { Outlet } from "react-router";
-import { Sidebar } from "@/components/layout/sidebar/Sidebar";
-import { HamburgerMenu } from "./sidebar/HamburgerMenu";
-import { useState } from "react";
+import { useEffect, useState } from "react";
+
+import { HamburgerMenu } from "@/components/navigation/HamburgerMenu";
+import { Sidebar } from "@/components/navigation/Sidebar";
+
+import { Checklist } from "@/types/Checklist";
+import { MOCK_CHECKLISTS } from "@/mock-data/mock-checklists";
+import { NavContext } from "@/Contexts";
 
 export default function TripLayout() {
   const [pageVisible, setPageVisible] = useState(true);
+  const [selected, setSelected] = useState("");
+  const [checklists, setChecklists] = useState<Checklist[]>([]);
+
+  // fetch checklists
+  useEffect(() => setChecklists(MOCK_CHECKLISTS), []);
 
   return (
-    <div className="flex flex-col md:flex-row">
-      <HamburgerMenu
-        setPageVisible={setPageVisible}
-        className="flex md:hidden"
-      />
-      <Sidebar className="hidden md:flex w-60" />
-      {pageVisible && (
-        <main className="grow flex justify-center">
-          <Outlet />
-        </main>
-      )}
-    </div>
+    <NavContext.Provider
+      value={{ selected, setSelected, checklists, setChecklists }}
+    >
+      <div className="flex flex-col md:flex-row">
+        <HamburgerMenu
+          setPageVisible={setPageVisible}
+          className="flex md:hidden"
+        />
+        <Sidebar className="hidden md:flex w-60" />
+        {pageVisible && (
+          <main className="grow flex justify-center">
+            <Outlet />
+          </main>
+        )}
+      </div>
+    </NavContext.Provider>
   );
 }
