+++ client/src/components/navigation/SidebarFooter.tsx
