+++ client/src/components/navigation/SidebarItem.tsx
